New Scene + can place turrets (doesn't work)
Created a new Maze-like level with the basic layout for now. Created a
new script to let the player place turrets on boxes, but doing so makes
the game crash because the turrets take in an enemy transform on
initialization. It needs to be able to find the enemy without us
providing it.